Rename DemoComponents to DocsComponents
